Stop loading google-chart-loader from google-chart
Remove dependency on google-chart-loader from google-chart and extract
shared loading code to a new module 'loader'.

Loading google-chart (the most common case) is now more efficient. The
code for google-chart-loader does not need be included in the bundle and
executed.
