+++ test/basic-tests.html
@@ -25,6 +25,7 @@
 <script type="module">
 import '@polymer/promise-polyfill/promise-polyfill-lite.js';
 import '../google-chart.js';
+import '../google-chart-loader.js';
 suite('<google-chart>', function() {
   var chart;
   setup(function() {
