+++ test/custom-load-test.html
@@ -19,7 +19,7 @@
 </test-fixture>
 
 <script type="module">
-  import {load} from '../google-chart-loader.js';
+  import {load} from '../loader.js';
   import '../google-chart.js';
 
   // This test has to run separately because Google Charts API is loaded only once per document.
